Convert numerical dtypes from strings to real types.
This only covers the simple numerical types. That is, types that have an
explicit byte order are unchanged for the moment. String types are also
unchanged.

See #823.

+++ obspy/realtime/tests/test_signal.py
@@ -27,7 +27,8 @@ class RealTimeSignalTestCase(unittest.TestCase):
         super(RealTimeSignalTestCase, self).__init__(*args, **kwargs)
         # read test data as float64
         self.orig_trace = read(os.path.join(os.path.dirname(__file__), 'data',
-                                            'II.TLY.BHZ.SAC'), dtype='f8')[0]
+                                            'II.TLY.BHZ.SAC'),
+                               dtype=np.float64)[0]
         # make really sure test data is float64
         self.orig_trace.data = np.require(self.orig_trace.data, 'f8')
         self.orig_trace_chunks = self.orig_trace / NUM_PACKETS
