+++ obspy/wav/core.py
@@ -150,7 +150,8 @@ def writeWAV(stream, filename, framerate=7000, rescale=False, width=4,
             if rescale:
                 # optimal scale, account for +/- and the zero
                 maxint = 2 ** (width * 8 - 1) - 1
-                data = data.astype('f8')  # upcast for following rescaling
+                # upcast for following rescaling
+                data = data.astype(np.float64)
                 data = data / abs(data).max() * maxint
             data = np.require(data, dtype=dtype)
             w.writeframes(data.tostring())
